Completes ex 12.7, adds BE :id routes for get&put
This commit implements GET and PUT routes for todos/:id. It also
incorporates updates to the docker-compose file, and adds a new
docker-compose.dev file to run a mongo server in docker while running
the backend locally for dvelopment. These changes to docker files are
dictated in the course material, and are not part of the exercise.